refactor(http): refactor http client
Refs #3481
